extracted case studies comp
